Refactor custom.js
