+++ src/___scripts/index.js
@@ -1,0 +1,2 @@
+import './checkout'
+import './handler'
