--- src/vendors.js
@@ -1,7 +1,0 @@
-// import node_modules vendors
-const vendors = () => {
-  console.log('vendors')
-  return 'vendors'
-}
-
-export default vendors
