+++ webpack/dev.config.js
@@ -2,17 +2,26 @@ const path = require('path')
 const HtmlWebpackPlugin = require('html-webpack-plugin')
 const MiniCssExtractPlugin = require('mini-css-extract-plugin')
 const CleanWebpackPlugin = require('clean-webpack-plugin')
+const { name } = require('../package.json')
+
+const pathsToClean = ['dist']
+
+const cleanOptions = {
+  root: path.join(__dirname, '../'),
+  exclude: ['img'],
+  verbose: true,
+  dry: false
+}
 
 module.exports = {
   mode: 'development',
   devtool: 'source-map',
   entry: {
-    app: './src/index.js',
-    vendors: './src/vendors.js'
+    app: './src/index.js'
   },
   output: {
     path: path.join(__dirname, '../', 'dist'),
-    filename: `[name].js`,
+    filename: path.join('js', `${name}.js`),
     publicPath: '/dist/'
   },
   module: {
@@ -28,19 +37,14 @@ module.exports = {
       {
         test: /\.js$/,
         exclude: /node_modules/,
-        use: {
-          loader: 'babel-loader',
-          options: {
-            presets: ['@babel/preset-env']
-          }
-        }
+        use: { loader: 'babel-loader', options: { presets: ['@babel/preset-env'] } }
       }
     ]
   },
   plugins: [
-    new CleanWebpackPlugin(['dist']),
+    new CleanWebpackPlugin(pathsToClean, cleanOptions),
     new MiniCssExtractPlugin({
-      filename: 'style.css'
+      filename: path.join('css', `${name}.css`)
     }),
     new HtmlWebpackPlugin({
       filename: 'index.html',
